refactor: update locks
