+++ commons-model/src/test/java/com/xenoterracide/commons/model/test/NameableTest.java
@@ -1,7 +1,7 @@
 // © Copyright 2024 Caleb Cushing
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-package com.xenoterracide.tm.model;
+package com.xenoterracide.commons.model.test;
 
 import static org.assertj.core.api.Assertions.assertThat;
